// resources/js/api/generated/models/imageInfo.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */

export interface ImageInfo {
  created: string;
  dangling: boolean;
  id: string;
  repository: string;
  size: number;
  tag: string;
  unused: boolean;
}